+++ apps/modernization-ui/src/apps/report/run/filters/advanced/ValueInput.tsx
@@ -1,4 +1,4 @@
-import React, { useId, useRef, useLayoutEffect } from 'react';
+import React, { useId, useEffect } from 'react';
 import { FullField, ValueEditorProps } from 'react-querybuilder';
 import { NumericInput, TextInputField } from '../../../../../design-system/input';
 import { DatePickerInput } from '../../../../../design-system/date';
@@ -6,6 +6,7 @@ import { DatePickerRange } from '../../../../../design-system/date/range/DatePic
 import { DateBetweenCriteria } from '../../../../../design-system/date/criteria';
 import { NumberInputRange } from '../../../../../design-system/input/range/NumberInputRange.tsx';
 import { NumberBetweenCriteria } from '../../../../../design-system/input/range/NumberRangeField.tsx';
+import { BETWEEN_OPERATOR } from './operators.ts';
 
 const RANGE_COMPONENTS = {
     date: DatePickerRange,
@@ -18,9 +19,11 @@ const SINGLE_COMPONENTS = {
     text: TextInputField,
 } as const;
 
+const BETWEEN_OPERATOR_NAME = BETWEEN_OPERATOR.name;
+
 const getConvertedRange = (props): DateBetweenCriteria | NumberBetweenCriteria => {
-    if (props.operator === 'between' && typeof props.value === 'string' && props.value) {
-        if (props.operator === 'between' && props.value) {
+    if (props.operator === BETWEEN_OPERATOR_NAME && typeof props.value === 'string' && props.value) {
+        if (props.operator === BETWEEN_OPERATOR_NAME && props.value) {
             const [from = '', to = ''] = props.value.split(',');
             return { between: { from, to } };
         }
@@ -30,26 +33,21 @@ const getConvertedRange = (props): DateBetweenCriteria | NumberBetweenCriteria =
 
 const ValueInput = (props: ValueEditorProps<FullField>) => {
     const id = useId();
-    const labelName = props.title ?? '';
-    const isBetween = props.operator === 'between';
-    const InputComponent = isBetween ? RANGE_COMPONENTS[props.inputType] : SINGLE_COMPONENTS[props.inputType];
+    const { handleOnChange, inputType, operator, title, value } = props;
+    const labelName = title ?? '';
+    const isBetween = operator === BETWEEN_OPERATOR_NAME;
+    const InputComponent = isBetween ? RANGE_COMPONENTS[inputType] : SINGLE_COMPONENTS[inputType];
 
-    let value = isBetween ? getConvertedRange(props) : (props.value ?? '');
+    let convertedValue = isBetween ? getConvertedRange(props) : (value ?? '');
 
-    const onChangeRef = useRef(props.handleOnChange);
-
-    useLayoutEffect(() => {
-        onChangeRef.current = props.handleOnChange;
-    }, [props.handleOnChange]);
-
-    // clean up value when value input is removed from DOM
-    useLayoutEffect(() => {
-        return () => {
-            if (!props.value) {
-                onChangeRef.current('');
-            }
-        };
-    }, []);
+    // eslint-disable-next-line max-len
+    // adapated from https://github.com/mcmcgrath13/react-querybuilder/blob/87a991b124fa9060431ac8e1e8f42b789a5ddecb/packages/react-querybuilder/src/components/ValueEditor.tsx#L312-L322
+    useEffect(() => {
+        // clear input when changing from between to other operators
+        if (![BETWEEN_OPERATOR_NAME].includes(operator) && value.includes(',')) {
+            handleOnChange('');
+        }
+    }, [handleOnChange, operator, value]);
 
     const handleSingleOnChange = (newValue: number | string) => {
         props.handleOnChange(newValue.toString());
@@ -71,7 +69,7 @@ const ValueInput = (props: ValueEditorProps<FullField>) => {
             <InputComponent
                 id={id}
                 label={isBetween ? '' : labelName}
-                value={value}
+                value={convertedValue}
                 name={labelName}
                 onChange={isBetween ? handleBetweenOnChange : handleSingleOnChange}
                 required
