+++ apps/modernization-ui/src/apps/report/run/filters/advanced/advanced-filter.module.scss
@@ -3,8 +3,6 @@
 @use 'styles/colors';
 
 .layout {
-    max-width: 90rem;
-
     :global {
         .queryBuilder {
             padding: 1rem;
